+++ CHANGELOG.md
@@ -28,8 +28,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   * Anchor (`a`) and related attributes (`href`, `target` and `download`)
   * Image (`img`) and related attributes (`src`, `alt`, `width` and `height`)
   * Lists (`ul`, `ol` and `li`)
-  * Forms (`form`, `input`) and related attributes
-    (`action`, `method`, `placeholder`, `for`, `value`, `min`, `max`, `minlength`, `maxlength`, `multiple`)
+  * Forms (`form`, `input`, `label`, `select`, `option`) and related attributes
+    (`action`, `method`, `placeholder`, `for`, `value`, `min`, `max`, `minlength`, `maxlength`, `multiple`, `type_*`)
   * Escape hatches (`raw` and `raw_unsafe`)
 * implement `From<(&'static str, Cow<'static, str)>` for `Attribute`
 * implement `From<&'static str>` and `From<String>` for `Element`